% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawWorldMapAll.R
\name{drawWorldMapAll}
\alias{drawWorldMapAll}
\title{Draw summary of all tambora data to a map}
\usage{
drawWorldMapAll(tamboraData)
}
\arguments{
\item{tamboraData, }{i.e. as inquired by function fromTambora()}
}
\description{
This function allows you to draw data inquired from tambora.org to a map
}
\examples{
data<-fromTambora(); drawWorldMapAll(data)
}
\keyword{map}
\keyword{tambora.org}
\keyword{visualisation}

