% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromTambora.R
\name{fromTambora}
\alias{fromTambora}
\title{A tambora.org inquire function}
\usage{
fromTambora(searchString, userEmail, userToken, baseUrl)
}
\arguments{
\item{searchString}{defines the properties to search for, i.e. 'g[cid]=58'}

\item{userEmail}{defines the email of the user account to use permissions}

\item{userToken}{defines the secret token of the user account to use permissions}

\item{baseUrl}{defines the instance of tambora to search from Defaults to 'https://www.tambora.org/index.php/'}
}
\description{
This function allows you to fetch data from tambora.org
}
\examples{
fromTambora()
}
\keyword{api}
\keyword{climatology}
\keyword{cre}
\keyword{tambora}

