% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmbImportPackages.R
\name{tmbImportPackages}
\alias{tmbImportPackages}
\title{Install and load multiple packages}
\usage{
tmbImportPackages(pkg)
}
\arguments{
\item{pkg}{list of packages, i.e.   c("jsonlite", "httr")}
}
\description{
This function allows you to install and load multiple packages
}
\examples{
tmbImportPackages(c("jsonlite", "httr"))
}
\keyword{install}
\keyword{packages}

